refactor: move admin kpis to updates hub
